Fix Release Build Environment
+ addressed every code safety demands compiler had in order to make it possible to build in release configuration without suppressing warning(warning C4244 still had to be suppressed because it was coming from external libraries xxhash and stb)

+ changed to proxying XInput9_1_0 instead of directx11 interface

+ recoded protocolbuffer helper

+ improved variable namings in fileshare

- removed ida_defs.h header as its not needed anymore

+++ source/proxy-dll/component/console.cpp
@@ -13,7 +13,7 @@ namespace console
 	namespace
 	{
 		const char* branding_str = "shield> ";
-		size_t branding_length = std::strlen(branding_str);
+		int branding_length = static_cast<int>(std::strlen(branding_str));
 
 		utilities::hook::detour printf_hook;
 		std::recursive_mutex print_mutex;
